update testing for new pension module
